+++ algorithm/short_lived_fake_kamat/FakeMessage.h
@@ -11,7 +11,7 @@ typedef nx_struct FakeMessage {
 
 } FakeMessage;
 
-inline int32_t Fake_get_sequence_number(const FakeMessage* msg) { return msg->sequence_number; }
+inline int64_t Fake_get_sequence_number(const FakeMessage* msg) { return msg->sequence_number; }
 inline int32_t Fake_get_source_id(const FakeMessage* msg) { return msg->source_id; }
 
 #endif // SLP_MESSAGES_FAKEMESSAGE_H
